requirements.txt
# Google Cloud Services
google-cloud-speech>=2.21.0
google-cloud-aiplatform>=1.38.0
google-cloud-texttospeech>=2.16.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
langdetect>=1.0.9

# Testing (optional)
aiohttp>=3.9.0

